Move tile lay when company sold logic to a module and share amongst 1889 and 1828.
