+++ lib/engine/step/g_1828/special_track.rb
@@ -2,12 +2,14 @@
 
 require_relative '../special_track'
 require_relative 'acquire_va_tunnel_coal_marker'
+require_relative '../tile_lay_when_company_sold'
 
 module Engine
   module Step
     module G1828
       class SpecialTrack < SpecialTrack
         include AcquireVaTunnelCoalMarker
+        include TileLayWhenCompanySold
       end
     end
   end
